Select environmental covariables based on phenotype data

Perform variable selection using observed phenotype data to determine the most relevant set of environmental covariables

ec_variable_selection(
  pheno.data,
  env.data,
  ec.matrix,
  env.col,
  gen.col,
  y.col,
  ec.subset = NULL,
  search.method = c("stepwise.lmm", "lasso"),
  cv.type = c("loo", "mc"),
  cv.ptrain = 0.6,
  cv.iter = 25,
  ...
)

ec_variable_selection_slide(
  pheno.data,
  env.data,
  env.col,
  gen.col,
  y.col,
  var.col,
  val.col,
  ec.subset = NULL,
  search.method = c("stepwise.lmm", "lasso"),
  time.col,
  time.list,
  cv.type = c("loo", "mc"),
  cv.ptrain = 0.6,
  cv.iter = 25
)

Arguments

pheno.data
A data.frame of phenotype data that will be used in the analysis or to train predictive models. This must contain columns for genotypes, environments, and the value of the phenotypic observation.

env.data
A data.frame of environmental variables in long format. Must be provided if ec.matrix is not passed; ignored otherwise. Must include information on the environments/trials/location in pheno.data.

ec.matrix
A matrix of (ideally) centered and scaled environmental covariable data (i.e. from the ec_matrix function. Supercedes env.data if passed. Must include information on the environments/trials/location in pheno.data.

env.col
The column in pheno.data and (if passed) env.data that designates environments.

gen.col
The column in pheno.data that designates genotypes.

y.col
The column in pheno.data that designates phenotypic values.

ec.subset
A character vector of covariables in env.data or ec.matrix that should be considered. If NULL (default), all covariables in env.data or ec.matrix are considered.

search.method
The covariable selection search method. See Details for more information.

cv.type
The type of cross-validation used to select variables. Can be "loo" for leave-one-out (environments) or "mc" for random train-test partitions.

cv.ptrain
The proportion of data to use in model training if cv.type = "mc".

cv.iter
The number of cross-validation replications if cv.type = "mc".

...
Other arguments to pass to the ec_matrix function (if ec.matrix is not passed).

var.col
See ec_matrix.

val.col
See ec_matrix.

time.col
The column in env.data used to create time frames (this is usually year).

time.list
A list of vectors, each containing elements of time.col that delineate the timeframes. If not passed, a list is created that accumulates elements of time.col from maximum to minimum in steps of 1 unit.

Value

The function ec_variable_selection returns a list with the following elements:

optVariables

A n x 1 matrix where the row names are the n environmental covariables selected and the elements of the matrix are the importance value of that variable.

finalCVResults

A vector with the cross-validation results of the final model.

CVStepResults

A matrix detailing the cross-validation results at each step of the stepwise linear model variable selection approach (i.e. search.method = "stepwise.lmm")

The function ec_variable_selection_slide returns a list with the following elements:

timeframeFinalCVResults

A data.frame of cross-validation results for each timeframe.

variableSelectionResults

A list containing the same elements as the output of ec_variable_selection, but only for the optimal timeframe.

ecMatrixSelected

The n x p matrix of environmental covariables summarized across the optimal timeframe.

Details

TBD.

Examples

if (FALSE) {

pheno.data <- pheno_data_train
env.data <- env_data_merged

variable_selection_out <- ec_variable_selection(
  pheno.data = pheno.data, env.data = env.data, env.col = "trial",
  gen.col = "line_name", y.col = "value", var.col = "variable",
  val.col = "value"
)

}


if (FALSE) {

# Replace location
pheno.data <- merge(pheno_data_train[-2], trial_info[c("trial", "location")])
env.data <- hist_env_data_merged

variable_selection_out <- ec_variable_selection_slide(
  pheno.data = pheno.data, env.data = env.data, env.col = "location",
  gen.col = "line_name", y.col = "value", var.col = "variable",
  val.col = "value", time.col = "year"
)

}