Predict historical planting dates

Uses the crop calendar of long-term planting date information from the Center for Sustainability and the Global Environment at the University of Wisconsin to predict historical planting dates using historical weather data.

predict_historical_planting_dates(
  env.data,
  crop,
  crop.calendar,
  meant.adj = -2,
  daylen.adj = -0.5,
  mint.lower = 0,
  rain.upper = 0.5
)

Arguments

env.data
A data.frame of historical environmental data from the output of get_historical_weather_data.

crop
The type of crop for which planting date information is desired. Must be one of the crops listed in the crop calendar.

crop.calendar
The data.frame of crop calendar information (available here). If not passed, the function will read in the data from this url (requires an internet connection).

meant.adj
The amount to adjust the mean temperature at planting from the crop calendar when using this mean temperature as a minimum threshold to determine eligible planting dates from historical weather data.

daylen.adj
The amount to adjust the daylength at planting from the crop calendar when using this daylength as a minimum threshold to determine eligible planting dates from historical weather data.

mint.lower
The lowest daily minimum temperature acceptable for determining eligible planting dates from historical weather data.

rain.upper
The highest daily rainfall acceptable for determining eligible planting dates from historical weather data.

Details

References

http://nelson.wisc.edu/sage/data-and-models/crop-calendar-dataset/index.php

Sacks, W.J., D. Deryng, J.A. Foley, and N. Ramankutty (2010). Crop planting dates: an analysis of global patterns. Global Ecology and Biogeography 19, 607-620. DOI: 10.1111/j.1466-8238.2010.00551.x.

Examples


if (FALSE) {
# Find distinct location info
location_info <- dplyr::distinct(trial_info, location, latitude, longitude)

# Retrieve historical weather data from 2005-2014 (10 years)
location_historical_weather <- get_historical_weather_data(
  location.info = location_info, start.year = 2005,
  end.year = 2014, source = "daymet"
)

# Download the crop calendar data
crop_calendar <- read.csv("http://nelson.wisc.edu/sage/data-and-models/crop-calendar-dataset/All_data_with_climate.csv")

# Predict planting dates
location_predicted_planting <- predict_historical_planting_dates(
  env.data = location_historical_weather, crop = "barley - spring",
  crop.calendar = crop_calendar
)

}